Rewrite the project README
This commit provides a much-needed update to the `README.md` of
wit-bindgen. The README now mentions WIT, components, the current
tooling workflow, and an updated "tutorial" for languages like Rust and
C. Overall it's intended to generally be more useful and up-to-date.

This commit additionally updates to the latest ABI for `wit-component`
and fixes a few minor issues here and there with the CLI.