Remove the JS host generator from this repository (#463)
* Refactor root `Cargo.toml` a bit

* Update some wasm-tools dependencies
* Use a `[patch]` for Wasmtime to reflect how this will eventually be a
  crates.io dependency
* Use `workspace = true` to avoid duplicating path names.

* Remove an unneeded futures-util dep

* Remove the JS host generator

This commit removes the JS host generator from this repository as it now
lives at bytecodealliance/js-component-tools. The runtime tests will be
updated in a subsequent commit to use Wasmtime as a host for executing
components.

This additionally removes the demo as it's no longer possible to build
it in just this repository. The demo has slowly been getting less useful
over time as generators are removed from this repository. Additionally
it's always been a demo and not much more so it's more-or-less served
its purpose and doesn't seem to have a great place to fit any more. This
can possibly exist out-of-tree eventually in a separate location that
pulls together all known generators, but until that time it seems best
to just remove it for now.

Closes #395

* Update the CLI with just guest generators

No need for a `guest` subcommand any more since all supported generators
are just guest generators.

* Let tests run for the `wit-bindgen` binary

This'll make sure that it builds as part of `cargo test`

* Add scaffolding to run tests in Wasmtime

Tests will unfortunately take longer due to wasmtime build times being
somewhat nontrivial and additionally rebuilding the java components on
each invocation is costly, but for now this should keep the tests
relatively simple and easy to maintain by having all the logic be in one
place rather than across multiple crates.

* Migrate flavorful test

* Migrate the lists test

* Migrate the many arguments test

* Migrate the records test

* Migrate the smoke test

* Bring back the unions host test

* Migrate the variants test

* Migrate strings test to Wasmtime

* Remove invalid runtime test

This is a test for host generators which is no longer being tested in
this repository.

* Remove exports_only test

This is a JS host test primarily

* Fix a rustfmt error