Make the Rust bindings friendlier for ahead-of-time generation.
In place of `no_std`, add a `std_feature` option, which tells the code
generator to add `#[cfg(feature = "std")]` to any use of `std`. This
allows the same bindings support either std or no-std depending on the
cargo flags in the crate using them.

And, in place of `unchecked`, use `#[cfg(not(debug_assertion))]`, so
that we automatically disable checks in release builds.

This way we can have a bindings crate with one copy of the generated
output that works across std and no_std, and disables checks in release
builds.