Fix bindings generation with unused types in Rust
This commit fixes an issue where if a type was imported into an
interface but wasn't actually used anywhere then bindings weren't
correctly generated because inference about whether it was owned or
borrowed didn't run. The fix here is to skip generating bindings for
unused types since they're not actually needed anywhere anyway. If
necessary in the future bindings can be forcibly generated but for now
this is hopefully largely "just" fixing an edge case.