wit-bindgen-rust: link component type for imports.
This commit forces a reference to the component type in import sub-modules.

It fixes an issue when building bindings to a different crate and only
referencing an import binding.

As there was no reference to `__link_section`, the generated core module from
the LLVM toolchain was missing the component type custom section.

The fix simply performs what we're already doing for exports in the `export!`
macro, but adds the reference to each generated import submodule.