Ensure that structs and unions in C are never empty.
Empty structs and unions are a non-standard extension in C, and have a
different layout between C and C++, so avoid emitting them. If a generated
struct or union would have zero members, give it a `u8 __empty;` member so
that it's not actually empty, so that it gets a consistent layout no
matter how it's being compiled.