Update dependencies to latest.
This commit updates dependencies to latest.

It includes the latest wasm-tools, where support for resources has been stubbed
out with a `todo!`.

Hopefully that won't be too much of an annoyance for those actively developing
support for resources.

Additionally, `indexmap`, which is used extensively on many public APIs, has
been rev'd to a new major version; this resulted in no required source changes.