Adjustments to Rust guest code generation
This commit refactors the exact shape of the Rust guest code generated
by `wit-bindgen`. The original problem here was that `rustfmt` would
fail on the generated output due to extreme nesting and extreme line
lengths presumably causing `rustfmt`'s heuristics to give up. While the
issue has been fixed for the input WIT world this is a bit of a game of
whack-a-mole so there's not a great regression test suite to have here.

In general the changes here are:

* Where possible move library functionality into the guest crate which
  generated code calls into. For example removing branches with
  `cfg(debug_assertions)`.
* Trying to de-indent code by less using nested Rust expressions and
  moreso using `let foo = ...` and propagate `foo` as the result of an
  operation.
* Enum lifting operations are now helper functions to avoid generating a
  lift-per-callsite which can be quite a lot for errors.
* Various other minor adjustments as aliasing variant names as a local
  `V{N}` instead of `super::super::super::your::here::Type` or
  restructuring variant lifts/lowers to have less nesting and/or be
  easier for `rustfmt` to format.

I think there's still a good bit to improve here. For example the
generated code has a lot of blocks such as:

    for ... {
        let base = ...;
        {
            // use `base`
        }
    }

where the body of the `for` loop has an unnecessary block. I've left
fixing this for a future refactoring though.