Remove reliance on `wit_parser::abi`
This commit is the dual of bytecodealliance/wasm-tools#1149 where the
code is being moved from the `wit-parser` crate into this repository
where it's mainly developed. This commit adds a `wit_bindgen_core::abi`
module with the same contents as before with small refactorings to avoid
inherent methods on `Resolve`.

Note that this temporarily switches to a git dependency for the
wasm-tools family of crates which will be reverted back to a crates.io
dependency one they're published.