rust: Define `cabi_realloc` as a weak symbol
This commit updates the `wit-bindgen` Rust crate to define the
`cabi_realloc` symbol as a weak symbol. This is not easy because Rust
does not offer a stable means by which to do this. Despite this through
a combination of shell scripts and CI it should be possible to get this
all working together by using C to define a weak symbol calling a known
Rust symbol.

Closes #849