rust: Fix cases where `with` provides resources (#855)
* rust: Fix cases where `with` provides resources

This commit fixes a bug where if `with` was used to provide a resource
then the internal `resources` map wasn't updated correctly. This could
lead to "index not found" panics in #832. To fix this I've opted to
remove the `resources` map entirely as the `owned` field was already not
tracked and this additionally brings the Wasmtime and Rust generators a
bit closer together.

Instead now each `InterfaceId` has an entry of whether it was last seen
as an export or an import and this is consulted when determining whether
a resource is exported or not.

This fix then required shuffling around where the stub implementations
are generated to ensure that when the stubs for top-level exported
functions are generated the right metadata of "was this interface last
seen as an import" is generated correctly.

Closes #832

* Update crates/rust/src/lib.rs

Co-authored-by: Pat Hickey <pat@moreproductive.org>

---------

Co-authored-by: Pat Hickey <pat@moreproductive.org>