rust: Define `cabi_realloc` as a weak symbol (#851)
* rust: Define `cabi_realloc` as a weak symbol

This commit updates the `wit-bindgen` Rust crate to define the
`cabi_realloc` symbol as a weak symbol. This is not easy because Rust
does not offer a stable means by which to do this. Despite this through
a combination of shell scripts and CI it should be possible to get this
all working together by using C to define a weak symbol calling a known
Rust symbol.

Closes #849

* Fix typo in yml

* More yml changes

* Regenerate cabi_realloc files

* Strip custom sections

* Install wasm-tools on CI as well

* Pin to slightly older wasm-tools

* Fix native tests

* Subvert rate limiting