Preserve pointer provenance in the Rust backend. (#870)
* Preserve pointer provenance in the Rust backend.

Use the new `Pointer` and `Length` types in the Rust backend to emit
code that uses `*mut c_void` and `usize` instead of `i32` when working
with pointers and array lengths.

To represent `PointerOrI64`, use a `MaybeUninit<u64>`, since that type
can hold any `u64` and is documented to also preserve provenance.

This change happens to get the generated Rust code close to supporting
memory64, however it isn't complete; the abi code still emits hard-coded
`+ 4` offsets for loading the length of a pointer+length pair in memory.

* Fully-qualify `core` in macro expansions.

* Fix bitcasts between i32 and pointer.

* Implement Bitcast::I32ToP and PToI32 in more backends.

* Use `*mut u8` instead of `*mut c_void` for pointers.

And switch to `add` from `byte_add`.

* Add documentation to `Bitcast` arms about provenance.