C: Use pointer types
Teach the C backend to use C pointer types. This is similar to #870
for Rust. C is less concerned about provenance, but it makes the ABI
easier to understand and reduces the amount of casting needed in the
generated code.

While here, I also optimized the code for list deallocation slightly.