Start supporting the `wasm32-wasip2` target (#892)
This commit adds some initial support for `wasm32-wasip2` in
`wit-bindgen`. Everything works as-is but one of the benefits of a new
target is that we can move runtime bits such as `cabi_realloc` into the
standard library rather than in this crate. This PR sets up
infrastructure to test the `wasm32-wasip2` target and additionally
assumes rust-lang/rust#122411 for providing the implementation of
`cabi_realloc`.