test(*): hardening codegen tests by adding wasi proposals as submodules (#919)
* test(*): hardening codegen tests by adding wasi proposals as submodules

this commit adds a few git submodules for wasi proposals and placed them
in `tests/codegen` folder for testing purposes. it also changes the structure
now that it requires wit packages to have a `wit` folder inside the directory
similar to how wasi proposal structures.

Signed-off-by: Jiaxiao Zhou (Mossaka) <duibao55328@gmail.com>

* test(rust): skip wasi-http and wasi-cli

both have multiple worlds which require another refactoring of the
current test codebase to adopt that

Signed-off-by: Jiaxiao Zhou (Mossaka) <duibao55328@gmail.com>

* test(*): skip wasi-* tests on teavm and csharp

Signed-off-by: Jiaxiao Zhou (Mossaka) <duibao55328@gmail.com>

---------

Signed-off-by: Jiaxiao Zhou (Mossaka) <duibao55328@gmail.com>