add resource support to C# generator
This adds resource support to the C# generator and also fixes a few
miscellaneous issues that were preventing tests from passing.

I believe the code generated for imported resources is reasonably ergonomic,
although I'm pretty new to C#, so I'm open to suggestions to make it more
idiomatic.

As with other languages, the code for exported resources is a bit less
ergonomic, I'm using an `interface` to represent the API, but I also want
implementations to inherit code which manages the resource handle, for which I'm
currently using a `class`.  The upshot is that implementing an exported resource
involves both extending a `class` and implementing an `interface`.  Again, I'm
open to suggestions about improving this.  Note that exporting resources tends
to be a lot less common in end-user code than importing them, so we probably
don't need to obsess over ergonomics in that case.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

disable C# runtime tests on non-Windows platforms

As of this writing, they only work on Windows, although Linux support is on the
horizon.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

Update lib.rs

Fixes lots of warnings and makes the C# code more idiomatic

rename rep_table.cs to RepTable.cs; address PR feedback

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

per-resource RepTables; use zero as handle sentinel

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

add comments explaining handle mgmt for exported resources

Signed-off-by: Joel Dice <joel.dice@fermyon.com>