C#: fix warnings in generated code; add new CLI options
This adds two new options to the C# codegen CLI:

- `--internal`: make symbols `internal` rather than `public` by default
- `--skip-support-files`: skip generating the `cabi_realloc`, component type and similar files

Those are both useful for generating code that we plan to ship as part of the
.NET runtime.  In that case, we don't want any of the generated code to be
exposed as part of the public API, and we leave it to the application to supply
`cabi_realloc` and the component type.

The other changes address warnings and lints encountered when adding generated
code to the .NET runtime codebase, e.g. XML validation in doc comments, trailing
whitespace, redundant function signature modifiers, etc.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>