C#: use exceptions for functions returning `result` (#968)
* C#: use exceptions for functions returning `result`

This addresses part of #964.

For WIT functions returning `result<T, E>`, we now generate methods which return
`T` and throw `WitException` such that the `E` value may be retrieved using
`WitException.Value`.

I considered making `WitException` generic, but I couldn't find a precident for
generic exception classes when searching online, so I'm not sure how idiomatic
that would look like.  So, for now at least, you need to cast the value obtained
using `WitException.Value` to the expected type.  Happy to revisit this if
desired.

For WIT functions returning nested results, e.g. `result<result<T, E1>, E2>`, we
generate methods which return `T` and throw `WitException`, using
`WitException.NestingLevel` to disambiguate when `E1` and `E2` are the same.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* use 32-bit values for `results` test

This is a workaround for the `return pointer not aligned` issue, which is still
causing problems for Windows builds.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* update C# results test

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>