add component type WIT file to C# output
This is to avoid the need to add binary .o files (which are difficult to audit
from a security perspective) to the .NET runtime repository.  I will open
related PRs on the `wasm-tools` and `wasm-component-ld` repos to enable the
latter to accept WIT files at link (and componentization) time.

Note that I've taken this opportunity to remove the cabi_realloc C file from the
output, since we can rely on `wasi-libc` providing an implementation.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>