C#: do not drop resource handles in finalizers (#1040)
* C#: do not drop resource handles in finalizers

As of this writing, we cannot safely drop a handle to an imported resource from
a .NET finalizer because it may still have one or more open child resources.
Once WIT has explicit syntax for indicating parent/child relationships, we
should be able to use that information to keep track of child resources
automatically in generated code, at which point we'll be able to drop them in
the correct order from finalizers.

Fixes #1039

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* remove finalizers per PR feedback

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>